+++ src/main/java/site/scalarstudios/scalarpower/block/machine/wire/copper/CopperWireBlockEntity.java
@@ -1,6 +1,7 @@
 package site.scalarstudios.scalarpower.block.machine.wire.copper;
 
 import site.scalarstudios.scalarpower.block.machine.MachineUtils;
+import site.scalarstudios.scalarpower.block.machine.wire.BaseWireBlockEntity;
 import site.scalarstudios.scalarpower.power.NeoEnergyTransferUtil;
 import site.scalarstudios.scalarpower.block.ScalarPowerBlockEntities;
 import net.minecraft.core.BlockPos;
@@ -12,7 +13,7 @@ import net.minecraft.world.level.storage.ValueInput;
 import net.minecraft.world.level.storage.ValueOutput;
 import net.neoforged.neoforge.transfer.energy.EnergyHandler;
 
-public class CopperWireBlockEntity extends BlockEntity {
+public class CopperWireBlockEntity extends BaseWireBlockEntity {
     private static final int ENERGY_CAPACITY = MachineUtils.WIRE_BASE_CAPACITY;
     private static final int PUSH_PER_SIDE = MachineUtils.WIRE_BASE_THROUGHPUT;