+++ src/main/java/site/scalarstudios/scalarpower/item/custom/WrenchItem.java
@@ -1,0 +1,26 @@
+package site.scalarstudios.scalarpower.item.custom;
+
+import net.minecraft.ChatFormatting;
+import net.minecraft.network.chat.Component;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.TooltipFlag;
+import net.minecraft.world.item.component.TooltipDisplay;
+
+import java.util.function.Consumer;
+
+public class WrenchItem extends Item {
+    public WrenchItem(Properties properties) {
+        super(properties);
+    }
+
+    @SuppressWarnings("deprecated")
+    @Override
+    public void appendHoverText(ItemStack itemStack, TooltipContext context, TooltipDisplay display, Consumer<Component> builder, TooltipFlag tooltipFlag) {
+        super.appendHoverText(itemStack, context, display, builder, tooltipFlag);
+        builder.accept(Component.translatable("tooltip.scalarpower.wrench_configure").withStyle(ChatFormatting.GRAY));
+        builder.accept(Component.translatable("tooltip.scalarpower.wrench_cycle").withStyle(ChatFormatting.GRAY));
+    }
+}
+
+